Add files via upload
The Gaussian Mixture Models (GMM) approach is an unsupervised
learning algorithm because we don't know any values for the target feature.
Since it may be used to identify clusters in the data, the GMM is additionally
included under clustering algorithms. The aim of the GMM is to find these
clusters, and since we only have the number of clusters as information, the
GMM is an unsupervised approach. In order to do that, we attempt to fit our
dataset using a gaussian mixture. That is, we look for a variety of gaussian
distributions that might be utilized to explain the form of our dataset. It's
important to note that these Gaussian-shaped clusters need not be circular, as
they would be in the KNN technique, but rather can take any shape that a
multivariate Gaussian distribution can. In other words, a circle can only
fluctuate in diameter, whereas a GMM model can simulate any ellipsoid shape
due to its covariance matrix.